+++ NEWS.md
@@ -1,3 +1,5 @@
+# fcaR (development version)
+
 # fcaR 1.2.1
 
 Enhancements: